Rename model User to Client

+++ src/main/java/ru/slavaievlev/db/repository/IAreaRepository.java
@@ -3,5 +3,5 @@ package ru.slavaievlev.db.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.slavaievlev.db.model.Area;
 
-public interface IAreaRepository extends JpaRepository<Area, Integer> {
+public interface IAreaRepository extends JpaRepository<Area, Long> {
 }
