+++ src/main/java/ru/slavaievlev/db/repository/ICommentRepository.java
@@ -3,5 +3,5 @@ package ru.slavaievlev.db.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.slavaievlev.db.model.Comment;
 
-public interface ICommentRepository extends JpaRepository<Comment, Integer> {
+public interface ICommentRepository extends JpaRepository<Comment, Long> {
 }
