+++ src/main/java/ru/slavaievlev/db/repository/IPlaceRepository.java
@@ -3,5 +3,5 @@ package ru.slavaievlev.db.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.slavaievlev.db.model.Place;
 
-public interface IPlaceRepository extends JpaRepository<Place, Integer> {
+public interface IPlaceRepository extends JpaRepository<Place, Long> {
 }
