+++ src/main/java/ru/slavaievlev/db/repository/IStreetRepository.java
@@ -3,5 +3,5 @@ package ru.slavaievlev.db.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.slavaievlev.db.model.Street;
 
-public interface IStreetRepository extends JpaRepository<Street, Integer> {
+public interface IStreetRepository extends JpaRepository<Street, Long> {
 }
